fix: apply body theme class outside network aware layout
