+++ packages/curve-ui-kit/src/hooks/useLayoutStoreResponsive.ts
@@ -1,23 +1,14 @@
 import { useCallback, useEffect } from 'react'
 import { getPageWidthClassName, useLayoutStore } from '@ui-kit/features/layout'
-import { useUserProfileStore } from '@ui-kit/features/user-profile'
 
 export const useLayoutStoreResponsive = () => {
   const { document } = typeof window === 'undefined' ? {} : window
-  const theme = useUserProfileStore(state => state.theme)
-  const pageWidth = useLayoutStore(state => state.pageWidth)
   const setLayoutWidth = useLayoutStore(state => state.setLayoutWidth)
   const setPageVisible = useLayoutStore(state => state.setPageVisible)
 
   const handleResizeListener = useCallback(() => {
     if (window?.innerWidth) setLayoutWidth(getPageWidthClassName(window.innerWidth))
   }, [setLayoutWidth])
-
-  useEffect(() => {
-    if (!pageWidth || !document) return
-    document.body.className = `theme-${theme} ${pageWidth}`.replace(/ +(?= )/g, '').trim()
-    document.body.setAttribute('data-theme', theme)
-  }, [document, pageWidth, theme])
 
   useEffect(() => {
     if (!window || !document) return
@@ -27,11 +18,11 @@ export const useLayoutStoreResponsive = () => {
     handleVisibilityChange()
 
     document.addEventListener('visibilitychange', handleVisibilityChange)
-    window.addEventListener('resize', () => handleResizeListener())
+    window.addEventListener('resize', handleResizeListener)
 
     return () => {
       document.removeEventListener('visibilitychange', handleVisibilityChange)
-      window.removeEventListener('resize', () => handleResizeListener())
+      window.removeEventListener('resize', handleResizeListener)
     }
   }, [document, handleResizeListener, setPageVisible])
 }
